Added food price graph
Also, can use space to fastforward.

+++ GameViews.cpp
@@ -63,7 +63,7 @@ void GameViews::resizeCheck()
 	view->setSize(window->getSize().x, window->getSize().y);
 	DEFAULT_VIEW->setSize(window->getSize().x, window->getSize().y);
 
-	DEFAULT_VIEW->setCenter(getWindowWidth() / 2, getWindowHeight() / 2);
+	DEFAULT_VIEW->setCenter(double(getWindowWidth()) / 2.0, double(getWindowHeight()) / 2.0);
 
     //static double zoomLevel = 1;
     //DEFAULT_VIEW->zoom(zoomLevel); 
@@ -146,7 +146,7 @@ void GameViews::scroll(const sf::Time& clock)
     }
   if (keys[17])		// What is this? //This is r, so hitting r resets the view. Making make it shift R, or remove it. 
     {
-	view->setCenter(getWindowWidth()/2,getWindowHeight()/2); 
+	  view->setCenter(double(getWindowWidth()) / 2.0, double(getWindowHeight()) / 2.0);
       window->setView(*view);
     }
